update .gitignore.. new course classes

+++ src/main/java/com/java/lopputyo/model/Student.java
@@ -50,7 +50,7 @@ public class Student implements Serializable {
     }
 
     
-    public int getStudentId() {
+    public long getStudentId() {
         return this.studentId;
     }
     public void setCourse(Integer courseId){
